// src/typings/bedrock/schemas/blocks/blocks.d.ts
export type MaterialInstance =
  | string
  | {
      ambient_occlusion?: AmbientOcclusion;
      face_dimming?: FaceDimming;
      render_method?: RenderMethod;
      texture?: Texture;
    };
/**
 * Should this material have ambient occlusion applied when lighting? If true, shadows will be created around and underneath the block.
 */
export type AmbientOcclusion = boolean;
/**
 * Should this material be dimmed by the direction it's facing?
 */
export type FaceDimming = boolean;
// ...
export type RenderMethod = "opaque" | "double_sided" | "blend" | "alpha_test";
/**
 * Texture name for the material.
 */
export type Texture = string;